+++ examples/sandbox/sandbox.py
@@ -29,13 +29,14 @@ simConfig = specs.SimConfig()  # dictionary to store sets of simulation configur
 ###############################################################################
 
 netParams.scaleConnWeightModels = {'HH': 1.0}
-netParams.shape = 'cylinder'
+netParams.shape = 'ellipsoid'# 'cylinder' 
 netParams.sizeX = 100
+netParams.sizeY = 300
 netParams.sizeZ = 50
 
 
 # Population parameters
-netParams.addPopParams('PYR1', {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 1000}) # add dict with params for this pop 
+netParams.addPopParams('PYR1', {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 1000, 'xRange': [20,80], 'ynormRange': [0.3,0.7]}) # add dict with params for this pop 
 #netParams.addPopParams('PYR2', {'cellModel': 'HH', 'cellType': 'PYR', 'numCells': 1}) # add dict with params for this pop 
 netParams.addPopParams('background', {'cellModel': 'NetStim', 'rate': 20, 'noise': 0.5, 'start': 1, 'seed': 2})  # background inputs
 
@@ -98,6 +99,7 @@ simConfig.savePickle = 0 # Whether or not to write spikes etc. to a .mat file
 # simConfig.addAnalysis('plotRaster', {'spikeHist': 'subplot'})
 # simConfig.addAnalysis('plotTraces', {'include': [0,1]})
 simConfig.addAnalysis('plot2Dnet', {'view': 'xz'})
+
 
 
 ###############################################################################